requirements.txt
#
# requirements.txt  Andrew Belles  Jan 20th, 2026 
# 
# Standard python dependencies for if nix-shell is unavailable by user 
# 
# 
# 

# Scientific Computing Depedencies 
numpy
scipy 
pandas
scikit-learn

# Image/Plotting Depedencies
matplotlib 
PyYAML 
imageio 

# Geography Dependencies 
geopandas
rasterio
fiona
rasterstats
shapely
pyarrow

# Machine Learning Libraries 
libpysal 
optuna
umap-learn
networkit 
xgboost 
torch 
torchvision
torch-geometric
torch-scatter
torch-sparse

# manifold/objectives.py imports torchsort directly. On some systems it may need
# a source build against the local torch install instead of a prebuilt wheel.
torchsort
